+++ qrgui-core/qrgui-services/src/main/java/it/cambi/qrgui/services/emia/api/ITemi14Service.java
@@ -9,14 +9,13 @@ import it.cambi.qrgui.util.wrappedResponse.XWrappedResponse;
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
-public interface ITemi14Service<T>
-{
-    XWrappedResponse<Temi14UteCat, List<TreeNode<T, Long>>> findAll(HttpServletRequest sr, Temi14UteCatId id) throws NoCategoriesAllowedException;
+public interface ITemi14Service<T> {
+  XWrappedResponse<Temi14UteCat, List<TreeNode<T, Integer>>> findAll(
+      HttpServletRequest sr, Temi14UteCatId id) throws NoCategoriesAllowedException;
 
-    XWrappedResponse<Temi14UteCat, List<TreeNode<T, Long>>> saveCategory(HttpServletRequest sr, Temi14UteCat temi14)
-            throws NoCategoriesAllowedException;
+  XWrappedResponse<Temi14UteCat, List<TreeNode<T, Integer>>> saveCategory(
+      HttpServletRequest sr, Temi14UteCat temi14) throws NoCategoriesAllowedException;
 
-    XWrappedResponse<Temi14UteCat, List<TreeNode<T, Long>>> deleteCategory(HttpServletRequest sr, Temi14UteCat pk)
-            throws NoCategoriesAllowedException;
-
+  XWrappedResponse<Temi14UteCat, List<TreeNode<T, Integer>>> deleteCategory(
+      HttpServletRequest sr, Temi14UteCat pk) throws NoCategoriesAllowedException;
 }
